+++ docs/src/examples/components/Button/States/ButtonExampleDisabled.shorthand.tsx
@@ -6,6 +6,7 @@ const ButtonExampleDisabled = () => (
     <Flex gap="gap.smaller">
       <Button disabled content="Default" />
       <Button disabled content="Primary" primary />
+      <Button disabled secondaryAlt content="Secondary Alt" />
       <Button disabled icon="emoji" content="Click me" iconPosition="before" primary />
       <Button disabled circular icon="translation" />
       <Button